flask_service/requirements.txt
backoff==2.1.2
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
googleapis-common-protos==1.56.4
grpcio==1.49.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.12.0
opentelemetry-exporter-otlp-proto-grpc==1.13.0
opentelemetry-exporter-otlp-proto-http==1.13.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-requests==0.33b0
opentelemetry-instrumentation-wsgi==0.33b0
opentelemetry-proto==1.13.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
protobuf==3.20.2
requests==2.28.1
six==1.16.0
style==1.1.0
typing_extensions==4.3.0
update==0.0.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1